x86-64, libos: replace syscall instruction with SYSCALL_INST macro
This patch is a preparation so that syscall instruction can be
replaced easily with other instructions.
This patch doesn't change resulted instructions. Later patch
introduce such.

LibOS hooks system call and redirect the control to it so that
it can handle system call instead of kernel.
This patch make such change easier.